pyproject.toml
[tool.poetry]
name = "drf-homework"
version = "0.1.0"
description = ""
authors = ["Eugene Geym <geymzheka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
pillow = "^10.4.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}
black = "^24.4.2"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.26.0"
coverage = "^7.6.0"
drf-yasg = "^1.21.7"
setuptools = "^71.1.0"
stripe = "^10.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
